release 1.2

+++ src/controllers/score/score.controller.ts
@@ -1,7 +1,7 @@
-import { Body, Controller, Get, Header, Param, Post, Res, UseGuards, UseInterceptors } from '@nestjs/common';
+import { Body, Controller, Get, Header, Param, Post, Res, UseGuards, Request, UseInterceptors, Patch, Delete } from '@nestjs/common';
 import { FileFieldsInterceptor } from '@nestjs/platform-express';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
-import { getScoreDto } from 'src/dto/score/create-score.dto';
+import { DeleteScoreDto, EditScoreDto, getScoreDto } from 'src/dto/score/create-score.dto';
 import { upload } from 'src/services/common/common.service';
 import { ScoreService } from '../../services/score/score.service';
 import { Response } from 'express'
@@ -15,22 +15,9 @@ export class ScoreController {
   @Post()
   @UseInterceptors(
     FileFieldsInterceptor([{ name: 'file', maxCount: 1 }], upload))
-  importScore(@Body() request: any) {
+  importScore(@Request() req: any, @Body() request: any) {
     try {
-      return this.scoreService.importScore(request)
-    } catch (err: any) {
-      return {
-        statusCode: err.statuscode,
-        message: err.originalError
-      }
-    }
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get(':class_id')
-  getScoresByClass(@Param() param: getScoreDto) {
-    try {
-      return this.scoreService.getAllScoresByClassId(param.class_id)
+      return this.scoreService.importScore(req.user.userId, request)
     } catch (err: any) {
       return {
         statusCode: err.statuscode,
@@ -41,9 +28,9 @@ export class ScoreController {
 
   @UseGuards(JwtAuthGuard)
   @Get('toAnnounce/:class_id')
-  getScoresToAnnounceByClass(@Param() param: getScoreDto) {
+  getScoresToAnnounceByClass(@Request() req: any, @Param() param: getScoreDto) {
     try {
-      return this.scoreService.getScoresToAnnounceByClass(param.class_id)
+      return this.scoreService.getScoresToAnnounceByClass(req.user.userId, param.class_id)
     } catch (err: any) {
       return {
         statusCode: err.statuscode,
@@ -54,10 +41,14 @@ export class ScoreController {
 
   @UseGuards(JwtAuthGuard)
   @Get('template/:class_id')
-  @Header('Content-type', 'text/xlsx')
-  async getScoreTemplate(@Param() param: getScoreDto, @Res() res: Response) {
+  @Header('Content-type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
+  async getScoreTemplate(@Request() req: any, @Param() param: getScoreDto, @Res() res: Response) {
     try {
-      const result = await this.scoreService.getScoreTemplate(param.class_id)
+      const result: any = await this.scoreService.getScoreTemplate(req.user.userId, param.class_id)
+      if (result.statusCode) {
+        res.send(result)
+        return res
+      }
       res.download(`${result}`, function (err) {
         fs.unlinkSync(`./${result}`)
       })
@@ -65,6 +56,46 @@ export class ScoreController {
         statusCode: 200,
         message: "success"
       }
+
+    } catch (err: any) {
+      return {
+        statusCode: err.statuscode,
+        message: err.originalError
+      }
+    }
+  }
+
+  @UseGuards(JwtAuthGuard)
+  @Get(':class_id')
+  getScoresByClass(@Request() request: any, @Param() param: getScoreDto) {
+    try {
+      return this.scoreService.getAllScoresByClassId(request.user.email, param.class_id)
+    } catch (err: any) {
+      return {
+        statusCode: err.statuscode,
+        message: err.originalError
+      }
+    }
+  }
+
+  @UseGuards(JwtAuthGuard)
+  @Patch()
+  async editScore(@Request() req: any, @Body() body: EditScoreDto) {
+    try {
+      return await this.scoreService.editScoreByScoreIdStdId(req.user.userId, body)
+    } catch (err: any) {
+      return {
+        statusCode: err.statuscode,
+        message: err.originalError
+      }
+    }
+  }
+
+  @UseGuards(JwtAuthGuard)
+  @Delete(':score_id')
+  async deleteScoreByScoreId(@Request() req: any, @Param() param: DeleteScoreDto) {
+    try {
+      return await this.scoreService.deleteScoreByScoreId(param.score_id, req.user.userId)
     } catch (err: any) {
       return {
         statusCode: err.statuscode,
